requirements.txt
-r common/requirements.txt

oss2
selenium
uwsgi
cython
django<2
djangorestframework==3.9.0
maya
jsonfield
simplejson
pymysql
croniter
xlrd
django-cors-headers
djangorestframework-csv
pymysql
social-auth-app-django
django-jet
django-jsoneditor
pysocks
django-admin-view-permission
xmltodict
xlsxwriter
pandas
numpy
robobrowser
influxdb
pycountry
django-rest-swagger
drf-yasg==1.9.1
flex
swagger-spec-validator
kafka-python
websocket-client==0.53.0
futu-api
intriniorealtime
pika==0.12.0
lxml
tushare
oss2
aliyun-python-sdk-sts
celery==4.3.0
sentry-sdk==0.13.4
mootdx
matplotlib
tables